fix(core): clear AppleClang warnings in core library
macOS Apple clang surfaces a different warning set than the prior
Ubuntu/GCC pass. Fix the first-party core-library warnings at the root
cause (no pragmas, no -Wno):

- -Wpessimizing-move: drop redundant std::move on prvalue RandomMp<N>()
  returns (random.cpp)
- -Wunused-lambda-capture: drop unused [this] from Boost.Spirit parser
  semantic-action lambdas (settings_parsers/{algorithm,endgames,tracking})
- -Wsign-conversion: explicit static_cast at signed/unsigned boundaries
  (Eigen::Index for Eigen indexing, size_t for std container indexing)
- -Wunused-parameter / -Wunused-variable / -Wunused-but-set-variable:
  comment out or remove dead names/locals
- -Wreorder-adjacent / -Wmisleading-indentation: brace the brace-less if
  in SetVariables so the following statement is unambiguous
- -Wdeprecated-copy: declare EGBoundaryMetaData copy-assign = default

Core library now compiles warning-free; all 10 C++ test suites pass.

+++ core/include/bertini2/endgames/powerseries.hpp
@@ -414,7 +414,7 @@ public:
 		{			
 			using std::pow;
 
-			std::tie(s_times, s_derivatives) = TransformToSPlane(candidate, t0, num_pts, ContStart::Front);
+			std::tie(s_times, s_derivatives) = TransformToSPlane(static_cast<int>(candidate), t0, num_pts, ContStart::Front);
 			RealT cand_power{1/static_cast<RealT>(candidate)};
 			RealT curr_diff = (HermiteInterpolateAndSolve<ComplexT>(
 								  pow((most_recent_time-t0)/(times[0]-t0),cand_power), // the target time
@@ -545,7 +545,7 @@ public:
 		TimeCont<ComplexT> s_times;
 		SampCont<ComplexT> s_derivatives;
 
-		std::tie(s_times, s_derivatives) = TransformToSPlane(c, t0, num_pts, ContStart::Back);
+		std::tie(s_times, s_derivatives) = TransformToSPlane(static_cast<int>(c), t0, num_pts, ContStart::Back);
 		// the data was transformed to be on the interval [0 1] so we can hard-code the time-to-solve as 0 here.
 
 		Precision(result, Precision(s_derivatives.back()));
